1. 去除utils_cjson警告信息

+++ sdk/utils/misc/utils_cjson.c
@@ -44,7 +44,7 @@ static int cJSON_strcasecmp(const char *s1,const char *s2)
 	return tolower(*(const unsigned char *)s1) - tolower(*(const unsigned char *)s2);
 }
 
-static void *(*cJSON_malloc)(size_t sz) = HAL_Malloc;
+static void *(*cJSON_malloc)(uint32_t sz) = HAL_Malloc;
 static void (*cJSON_free)(void *ptr) = HAL_Free;
 
 static char* cJSON_strdup(const char* str)
